frontend/requirements.txt
astroid==2.4.1
autopep8==1.5.3
boto3==1.14.16
botocore==1.17.16
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cognitojwt==1.2.2
colorama==0.4.3
docutils==0.15.2
ecdsa==0.15
Flask==1.1.2
Flask-AWSCognito==1.3
flask-swagger==0.2.14
Flask-Mail==0.9.1   
idna==2.10
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pip-autoremove==0.9.1
pyasn1==0.4.8
pycodestyle==2.6.0
pylint==2.5.2
PyMySQL==0.9.3
python-dateutil==2.8.1
python-dotenv==0.13.0
python-jose==3.1.0
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
s3transfer==0.3.3
six==1.15.0
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
Werkzeug==1.0.1
wrapt==1.12.1
